- qCompany project now works on Qt5 (e.g. toAscii() methods converted to toLatin1() methods) - New version of QDecimal project with patches making it possible to work on Qt5 - qfakturyTests: Corrected input test data in CommodityDialogTest so that double numbers are always with proper decimal point (added locale awareness) - qfakturyTests: CommodityDialogPublic class is moved to separate header/class files - qfakturyTests: CommodityListDialogPublic class is moved to separate header/class files - qfakturyTests: CounterpartyDialogPublic class is moved to separate header/class files - qfakturyTests: Counterparty_t class is created for testing purposes - qfakturyTests: Added virtual users working in separate threads for handling application-blocking user dialogs: GuiUserAddCommodity, GuiUserAddCounterparty and GuiUserAddNewCommodity - qfakturyTests: InvoiceDialogPublic class is moved to separate header/ - qfakturyTests: Added new GUI unit tests in InvoiceDialogTest

+++ qfaktury/src/model/models/ModelCurrency.cpp
@@ -115,10 +115,10 @@ void ModelCurrency::updateDBsCurrenciesRates_()
 
     const QDomNodeList elements(xmlDoc.documentElement()
                                 .elementsByTagName("pozycja"));
-    for(size_t i = 0; i < elements.length(); ++i)
+    for(int i = 0; i < elements.length(); ++i)
     {
         const QDomNodeList children(elements.item(i).childNodes());
-        for(size_t j = 0; j < children.length(); ++j)
+        for(int j = 0; j < children.length(); ++j)
         {
             if(children.item(j).nodeName() == "nazwa_waluty")
             {
